💥 Upgrade to AWS JavaScript SDK V3 (#162)
Migration Guide:

When upgrading to this version, you also need to migrate from version 2 of the AWS JavaScript SDK to version 3. Instead of passing an instance of the old `aws.S3`, use the new `S3Client` class from `@aws-sdk/client-s3`.

The minimum required version of Node.js has also been increased to 12.0.0, to align with the AWS JavaScript SDK.